fix debug + convert score to double + remove example test + create test every levels solution + create test effective move

+++ android_app/app/src/main/java/com/slykos/bokoa/logic/models/Level.kt
@@ -2,6 +2,6 @@ package com.slykos.bokoa.logic.models
 
 class Level(
     @JvmField var operations: Array<Array<Operation>> = arrayOf(),
-    @JvmField var bestScore: Float = 0f,
+    @JvmField var bestScore: Double = 0.0,
     @JvmField var bestMoves: Array<String> = arrayOf()
 )
